backend/requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-simplejwt
django-jazzmin

